+++ lib/puppet/type/credential.rb
@@ -1,18 +1,20 @@
 require 'puppet/type'
 
+# Manages Credential Manager credentials on Windows systems.
 Puppet::Type.newtype(:credential) do
-  @doc = <<-EOT
-    Manages Credential Manager credentials on Windows systems.
-  EOT
+  desc "Manages Credential Manager credentials on Windows systems."
 
   ensurable
 
   newparam(:target, :namevar =>  true) do
+    desc "Conjur / DAP URL"
   end
 
   newproperty(:username) do
+    desc "The identity used to authenticate to the Conjur / DAP instance"
   end
 
   newproperty(:value) do
+    desc "The API key matching the Conjur / DAP identity"
   end
 end